flask-library/requirements.txt
# run with: pip install -r requirements.txt
flask
flask_cors
pymongo
dnspython
pyjwt